+++ kropki/src/main/java/com/kropki/bot/BotSolverImpl.java
@@ -2,7 +2,6 @@ package com.kropki.bot;
 
 import com.kropkigame.controller.GameBoardController;
 import com.kropkigame.view.Cell;
-import com.kropkigame.view.GameBoardPanel;
 import javafx.application.Platform;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
@@ -14,7 +13,6 @@ import javafx.scene.control.Alert.AlertType;
  */
 public class BotSolverImpl implements BotSolver {
     private GameBoardController gameBoardController;
-    private GameBoardPanel view;
     private boolean isRunning;
 
     /**
@@ -22,9 +20,9 @@ public class BotSolverImpl implements BotSolver {
      * @param gameBoardController Le contrôleur du plateau de jeu.
      * @param view La vue du plateau de jeu.
      */
-    public BotSolverImpl(GameBoardController gameBoardController, GameBoardPanel view) {
+    public BotSolverImpl(GameBoardController gameBoardController) {
         this.gameBoardController = gameBoardController;
-        this.view = view;
+        this.isRunning = false;
     }
 
     /**
@@ -56,7 +54,7 @@ public class BotSolverImpl implements BotSolver {
             gameBoardController.resetGame();
             int correctRow = 0;
             int correctCol = 0;
-            view.getCell(correctRow, correctCol).setNumber(gameBoardController.getModel().getNumber(correctRow, correctCol));
+            gameBoardController.getView().getCell(correctRow, correctCol).setNumber(gameBoardController.getModel().getNumber(correctRow, correctCol));
         });
     
         try {
@@ -69,12 +67,10 @@ public class BotSolverImpl implements BotSolver {
                         }
 
                         long endTime = System.currentTimeMillis();
-                        System.out.println("[BotSolver] Solution trouvée en " + (endTime - startTime) + " ms!");
                         showBotVictoryMessage(startTime, endTime);
                     }
                     stopBot(); // Arrête le bot si une solution est trouvée
                 } else {
-                    System.out.println("[BotSolver] Aucune solution trouvée.");
                     showNoSolutionFoundMessage();
                 }
             });
@@ -90,33 +86,27 @@ public class BotSolverImpl implements BotSolver {
      * @return true si une solution a été trouvée, false sinon.
      */
     private boolean solveWithBacktracking(int row, int col) {
-        System.out.println("[Backtracking] Cellule actuelle: [" + row + ", " + col + "]");
-
-        if (row == view.getGridSize()) { // Si toutes les lignes sont parcourues
+        if (row == gameBoardController.getView().getGridSize()) { // Si toutes les lignes sont parcourues
             return true;
         }
 
-        int nextRow = (col == view.getGridSize() - 1) ? row + 1 : row;
-        int nextCol = (col == view.getGridSize() - 1) ? 0 : col + 1;
+        int nextRow = (col == gameBoardController.getView().getGridSize() - 1) ? row + 1 : row;
+        int nextCol = (col == gameBoardController.getView().getGridSize() - 1) ? 0 : col + 1;
 
-        if (view.getCell(row, col).getNumber() != 0) {
-            System.out.println("[Backtracking] Cellule déjà remplie.");
+        if (gameBoardController.getView().getCell(row, col).getNumber() != 0) {
             return solveWithBacktracking(nextRow, nextCol);
         }
 
-        for (int num = 1; num <= view.getGridSize(); num++) {
-            System.out.println("[Backtracking] Essai de placement: " + num + " dans [" + row + ", " + col + "]");
+        for (int num = 1; num <= gameBoardController.getView().getGridSize(); num++) {
             if (isValidPlacement(row, col, num)) {
-                view.getCell(row, col).setNumber(num);
+                gameBoardController.getView().getCell(row, col).setNumber(num);
                 if (solveWithBacktracking(nextRow, nextCol)) {
                     return true;
                 }
-                System.out.println("[Backtracking] Retour en arrière, effacement de la cellule [" + row + ", " + col + "]");
-                view.getCell(row, col).setNumber(0); // Effacer et revenir en arrière
+                gameBoardController.getView().getCell(row, col).setNumber(0); // Effacer et revenir en arrière
             }
         }
 
-        System.out.println("[Backtracking] Aucun placement valide trouvé pour la cellule [" + row + ", " + col + "]");
         return false; // Aucun numéro valide trouvé pour cette cellule
     }
 
@@ -128,12 +118,10 @@ public class BotSolverImpl implements BotSolver {
      * @return true si le placement est valide, false sinon.
      */
     private boolean isValidPlacement(int row, int col, int num) {
-        System.out.println("[isValidPlacement] Vérification du placement de " + num + " en [" + row + ", " + col + "]");
     
         // Vérifier si le numéro est déjà dans la même ligne ou colonne
-        for (int i = 0; i < view.getGridSize(); i++) {
-            if (view.getCell(row, i).getNumber() == num || view.getCell(i, col).getNumber() == num) {
-                System.out.println("[isValidPlacement] Numéro " + num + " déjà présent dans la ligne ou la colonne");
+        for (int i = 0; i < gameBoardController.getView().getGridSize(); i++) {
+            if (gameBoardController.getView().getCell(row, i).getNumber() == num || gameBoardController.getView().getCell(i, col).getNumber() == num) {
                 return false;
             }
         }
@@ -144,16 +132,14 @@ public class BotSolverImpl implements BotSolver {
             int adjCol = col + direction[1];
     
             if (isValidCell(adjRow, adjCol)) {
-                Cell adjacentCell = view.getCell(adjRow, adjCol);
+                Cell adjacentCell = gameBoardController.getView().getCell(adjRow, adjCol);
                 int adjacentNumber = adjacentCell.getNumber();
-                System.out.println("[isValidPlacement] Cellule adjacente [" + adjRow + ", " + adjCol + "] = " + adjacentNumber);
 
                 // Vérifier les points noirs
                 if (gameBoardController.getModel().existsBlackEdgePoint(row + 1, col + 1, adjRow + 1, adjCol + 1)
                     || gameBoardController.getModel().existsBlackEdgePoint(adjRow + 1, adjCol + 1, row + 1, col + 1)) {
                     // Vérifie si la règle du point noir est respectée ou si la cellule adjacente est vide
                     if (adjacentNumber != 0 && !((num == 1 && adjacentNumber == 2) || (num == 2 && adjacentNumber == 1) || (num == 2 * adjacentNumber) || (adjacentNumber == 2 * num))) {
-                        System.out.println("[isValidPlacement] Règle du point noir non respectée pour " + num + " en [" + row + ", " + col + "]");
                         return false;
                     }
                 }
@@ -163,7 +149,6 @@ public class BotSolverImpl implements BotSolver {
                     || gameBoardController.getModel().existsWhiteEdgePoint(adjRow + 1, adjCol + 1, row + 1, col + 1)) {
                     // Vérifie si la règle du point blanc est respectée ou si la cellule adjacente est vide
                     if (adjacentNumber != 0 && Math.abs(num - adjacentNumber) != 1) {
-                        System.out.println("[isValidPlacement] Règle du point blanc non respectée pour " + num + " en [" + row + ", " + col + "]");
                         return false;
                     }
                 }
@@ -172,7 +157,6 @@ public class BotSolverImpl implements BotSolver {
                 if (!gameBoardController.getModel().existsBlackEdgePoint(row + 1, col + 1, adjRow + 1, adjCol + 1) && !gameBoardController.getModel().existsWhiteEdgePoint(row + 1, col + 1, adjRow + 1, adjCol + 1)
                     && !gameBoardController.getModel().existsBlackEdgePoint(adjRow + 1, adjCol + 1, row + 1, col + 1) && !gameBoardController.getModel().existsWhiteEdgePoint(adjRow + 1, adjCol + 1, row + 1, col + 1)) {
                     if (adjacentNumber != 0 && (Math.abs(num - adjacentNumber) == 1 || num == 2 * adjacentNumber || adjacentNumber == 2 * num)) {
-                        System.out.println("[isValidPlacement] Règle de l'absence de point non respectée pour " + num + " en [" + row + ", " + col + "]");
                         return false;
                     }
                 }
@@ -180,7 +164,6 @@ public class BotSolverImpl implements BotSolver {
             }
         }
     
-        System.out.println("[isValidPlacement] Placement valide pour " + num + " en [" + row + ", " + col + "]");
         return true; // Si toutes les vérifications sont passées, le placement est valide
     }
     
@@ -191,7 +174,7 @@ public class BotSolverImpl implements BotSolver {
      * @return true si la cellule est valide, false sinon.
      */
     private boolean isValidCell(int row, int col) {
-        return row >= 0 && row < view.getGridSize() && col >= 0 && col < view.getGridSize();
+        return row >= 0 && row < gameBoardController.getView().getGridSize() && col >= 0 && col < gameBoardController.getView().getGridSize();
     }
 
     /**
