+++ kropki/src/main/java/com/kropki/help/GameHelper.java
@@ -1,0 +1,13 @@
+package com.kropki.help;
+
+/**
+ * L'interface GameHelper définit les méthodes de base pour une aide à la résolution d'une grille de jeu.
+ * Elle est utilisée pour fournir une aide au joueur basée sur le respect des règles de Kropki.
+ */
+public interface GameHelper {
+    /**
+     * Fournit une aide en remplissant automatiquement certaines cases de la grille,
+     * en fonction des règles de Kropki.
+     */
+    void provideHelp();
+}
